Upload 5 files
This project is a News Summarization and Sentiment Analysis application, which fetches news articles related to a specified company and generates a sentiment analysis report. The app then provides a Text-to-Speech (TTS) functionality to read out the summary in Hindi. This is built using Streamlit for the user interface, utilizing Python libraries for sentiment analysis, text summarization, and TTS conversion.